refactor(native-rust): sync Esdk client extraction into src/client.rs from unreviewed
